Remove LEDnetWF BLE v2 integration files including light, number, and translation components, along with the manifest. This cleanup eliminates unused code and prepares for a potential refactor or new implementation.
