+++ custom_components/lednetwf_ble/capabilities.py
