+++ custom_components/lednetwf_ble/light.py
@@ -1,401 +1,403 @@
-import logging
-import voluptuous as vol
-from typing import Any, Optional
-from .lednetwf import LEDNETWFInstance
-from .const import (DOMAIN)
+"""Light platform for LEDnetWF BLE v2 integration."""
+from __future__ import annotations
 
-from homeassistant.const import CONF_MAC
-import homeassistant.helpers.config_validation as cv
-from homeassistant.helpers.entity import DeviceInfo
+import logging
+from typing import Any
+
 from homeassistant.components.light import (
-    PLATFORM_SCHEMA,
     ATTR_BRIGHTNESS,
     ATTR_COLOR_TEMP_KELVIN,
     ATTR_EFFECT,
-    EFFECT_OFF,
-    ATTR_HS_COLOR,
     ATTR_RGB_COLOR,
     ColorMode,
     LightEntity,
     LightEntityFeature,
 )
-from homeassistant.helpers import device_registry
+from homeassistant.config_entries import ConfigEntry
+from homeassistant.core import HomeAssistant, callback
+from homeassistant.helpers.device_registry import DeviceInfo
+from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-LOGGER = logging.getLogger(__name__)
-PLATFORM_SCHEMA = PLATFORM_SCHEMA.extend({vol.Required(CONF_MAC): cv.string})
+from .const import DOMAIN
+from .device import LEDNetWFDevice
 
-
-async def async_setup_entry(hass, config_entry, async_add_devices):
-    instance = hass.data[DOMAIN][config_entry.entry_id]
-    await instance.update()
-    
-    devices = [LEDNETWFLight(instance, config_entry.data["name"], config_entry.entry_id)]
-    
-    # Add background light entity only for models that support it (0x56, 0x80)
-    if hasattr(instance._model_interface, 'set_bg_color') and instance.model_number in [0x56, 0x80]:
-        LOGGER.info(f"Adding background light entity for model 0x{instance.model_number:02x}")
-        devices.append(LEDNETWFBackgroundLight(instance, config_entry.data["name"], config_entry.entry_id))
-    
-    async_add_devices(devices)
-    #config_entry.async_on_unload(await instance.stop())
+_LOGGER = logging.getLogger(__name__)
 
 
-class LEDNETWFLight(LightEntity):
-    _attr_has_entity_name = False
+async def async_setup_entry(
+    hass: HomeAssistant,
+    entry: ConfigEntry,
+    async_add_entities: AddEntitiesCallback,
+) -> None:
+    """Set up the light platform."""
+    device: LEDNetWFDevice = hass.data[DOMAIN][entry.entry_id]
 
-    def __init__(
-        self, lednetwfinstance: LEDNETWFInstance, name: str, entry_id: str
-    ) -> None:
-        self._instance = lednetwfinstance
-        self._entry_id = entry_id
-        self._attr_supported_color_modes = self._instance._model_interface.supported_color_modes
-        self._attr_supported_features = LightEntityFeature.EFFECT
-        self._attr_name               = name
-        self._attr_unique_id          = self._instance.mac
-        # Register this entity's callback
-        self._instance.register_callback(self.light_local_callback)
-        
-    @property
-    def available(self):
-        return self._instance.is_on is not None
+    entities: list[LightEntity] = [LEDNetWFLight(device, entry)]
+
+    # Add background light entity for devices that support it (0x56, 0x80)
+    if device.has_bg_color:
+        _LOGGER.info(
+            "Adding background light entity for %s (product_id=0x%02X)",
+            device.name,
+            device.product_id or 0,
+        )
+        entities.append(LEDNetWFBackgroundLight(device, entry))
+
+    async_add_entities(entities)
+
+
+class LEDNetWFLight(LightEntity):
+    """Representation of a LEDnetWF light."""
+
+    _attr_has_entity_name = True
+    _attr_name = None  # Use device name
+
+    def __init__(self, device: LEDNetWFDevice, entry: ConfigEntry) -> None:
+        """Initialize the light."""
+        self._device = device
+        self._entry = entry
+
+        # Set up unique ID
+        self._attr_unique_id = device.address
+
+        # Determine supported color modes
+        color_modes: set[ColorMode] = set()
+
+        if device.has_rgb:
+            color_modes.add(ColorMode.RGB)
+        if device.has_color_temp:
+            color_modes.add(ColorMode.COLOR_TEMP)
+
+        # If no color modes, at least support brightness
+        if not color_modes:
+            color_modes.add(ColorMode.BRIGHTNESS)
+
+        self._attr_supported_color_modes = color_modes
+
+        # Set up features
+        features = LightEntityFeature(0)
+        if device.has_effects:
+            features |= LightEntityFeature.EFFECT
+        self._attr_supported_features = features
+
+        # Color temp range
+        if device.has_color_temp:
+            self._attr_min_color_temp_kelvin = device.min_color_temp_kelvin
+            self._attr_max_color_temp_kelvin = device.max_color_temp_kelvin
+
+        # Register callback for state updates
+        device.register_callback(self._handle_state_update)
+
+    async def async_will_remove_from_hass(self) -> None:
+        """Handle entity removal."""
+        self._device.unregister_callback(self._handle_state_update)
+
+    @callback
+    def _handle_state_update(self) -> None:
+        """Handle state updates from the device."""
+        self.async_write_ha_state()
 
     @property
-    def brightness(self):
-        return self._instance.brightness
-    @property
-    def brightness_step_pct(self):
-        return 10
-    
-    @property
-    def is_on(self) -> Optional[bool]:
-        return self._instance.is_on
-
-    @property
-    def color_temp_kelvin(self):
-        return self._instance.color_temp_kelvin
-
-    @property
-    def max_color_temp_kelvin(self):
-        return self._instance.max_color_temp_kelvin
-
-    @property
-    def min_color_temp_kelvin(self):
-        return self._instance.min_color_temp_kelvin
-
-    @property
-    def effect_list(self):
-        return self._instance.effect_list
-
-    @property
-    def effect(self):
-        return self._instance.effect
-
-    @property
-    def supported_features(self) -> int:
-        """Flag supported features."""
-        return self._attr_supported_features
-
-    @property
-    def supported_color_modes(self) -> set[ColorMode]:
-        """Flag supported color modes."""
-        return self._attr_supported_color_modes
-
-    @property
-    def hs_color(self):
-        """Return the hs color value."""
-        return self._instance.hs_color
-  
-    @property
-    def rgb_color(self):
-        """Return the rgb color value."""
-        return self._instance.rgb_color
-
-    @property
-    def color_mode(self):
-        """Return the color mode of the light."""
-        return self._instance.color_mode
-    
-    @property
-    def firmware_version(self):
-        return self._instance.firmware_version
-    
-    @property
-    def device_info(self):
+    def device_info(self) -> DeviceInfo:
         """Return device info."""
+        # Model string: capability name from product ID mapping
+        cap_name = self._device.capabilities.get("name", "Unknown")
+        model_str = cap_name
+
+        # Software version: app-style firmware string
+        # Format: "{product_id:02X}.{firmware_ver:04d}.{ble_version:02d},V{led_version}"
+        # Example: "62.0008.05,V3" matching the LEDnetWF Android app display
+        sw_version = self._device.app_firmware_version
+
         return DeviceInfo(
-            identifiers={
-            (DOMAIN, self._instance.mac)
-            },
-            name=self.name,
-            connections={(device_registry.CONNECTION_NETWORK_MAC, self._instance.mac)},
-            model=f"0x{self._instance.model_number:02x}",
-            sw_version=self.firmware_version
+            identifiers={(DOMAIN, self._device.address)},
+            name=self._device.name,
+            manufacturer="LEDnetWF",
+            model=model_str,
+            sw_version=sw_version,
         )
 
     @property
-    def should_poll(self):
-        return False
+    def available(self) -> bool:
+        """Return True if entity is available."""
+        return self._device.is_on is not None
 
     @property
-    def name(self) -> str:
-        return self._attr_name
-    
+    def is_on(self) -> bool | None:
+        """Return True if light is on."""
+        return self._device.is_on
+
     @property
-    def icon(self):
-        return self._instance._model_interface.icon
-    
+    def brightness(self) -> int | None:
+        """Return the brightness."""
+        return self._device.brightness
+
+    @property
+    def rgb_color(self) -> tuple[int, int, int] | None:
+        """Return RGB color."""
+        return self._device.rgb_color
+
+    @property
+    def color_temp_kelvin(self) -> int | None:
+        """Return color temperature in Kelvin."""
+        return self._device.color_temp_kelvin
+
+    @property
+    def effect_list(self) -> list[str] | None:
+        """Return list of effects."""
+        effects = self._device.effect_list
+        return effects if effects else None
+
+    @property
+    def effect(self) -> str | None:
+        """Return current effect."""
+        return self._device.effect
+
+    @property
+    def extra_state_attributes(self) -> dict[str, Any]:
+        """Return extra state attributes for diagnostics."""
+        attrs: dict[str, Any] = {}
+
+        # Product ID (hex format for easy lookup in docs)
+        if self._device.product_id is not None:
+            attrs["product_id"] = f"0x{self._device.product_id:02X}"
+
+        # Effect type (command format used by device)
+        attrs["effect_type"] = self._device.effect_type.name
+
+        # Effect speed (when effect is active)
+        if self._device.effect:
+            attrs["effect_speed"] = self._device.effect_speed
+
+        # LED configuration (for addressable strips)
+        if self._device.led_count:
+            attrs["led_count"] = self._device.led_count
+        if self._device.segments:
+            attrs["segments"] = self._device.segments
+        if self._device.total_leds:
+            attrs["total_leds"] = self._device.total_leds
+
+        # Capabilities (useful for debugging)
+        attrs["has_rgb"] = self._device.has_rgb
+        attrs["has_color_temp"] = self._device.has_color_temp
+        if self._device.has_builtin_mic:
+            attrs["has_builtin_mic"] = True
+
+        # Device info from service data (protocol_docs/17_device_configuration.md)
+        if self._device.ble_version is not None:
+            attrs["ble_version"] = self._device.ble_version
+        if self._device.led_version is not None:
+            attrs["led_version"] = self._device.led_version
+        if self._device.firmware_ver is not None:
+            attrs["firmware_ver"] = self._device.firmware_ver
+        if self._device.firmware_flag is not None:
+            attrs["firmware_flag"] = f"0x{self._device.firmware_flag:02X}"
+
+        # Data-driven capabilities (from JSON database)
+        json_caps = self._device.json_capabilities
+        if json_caps:
+            attrs["json_device_version"] = self._device.device_version
+            attrs["json_state_protocol"] = json_caps.get_state_protocol(
+                self._device.device_version
+            )
+            # Show which effect functions are supported
+            attrs["supports_scene_data_v2"] = self._device.supports_datadriven_function(
+                "scene_data_v2"
+            )
+
+        return attrs
+
+    @property
+    def color_mode(self) -> ColorMode:
+        """Return current color mode."""
+        if self._device.effect:
+            # For Settled Mode effects, allow RGB color changes
+            # so user can adjust foreground color while staying in the effect
+            if self._device.is_in_settled_effect():
+                return ColorMode.RGB
+            # For other effects, report brightness mode (no color picker)
+            return ColorMode.BRIGHTNESS
+        if self._device.color_temp_kelvin and self._device.has_color_temp:
+            return ColorMode.COLOR_TEMP
+        if self._device.rgb_color and self._device.has_rgb:
+            return ColorMode.RGB
+        if ColorMode.BRIGHTNESS in self._attr_supported_color_modes:
+            return ColorMode.BRIGHTNESS
+        # Fallback
+        return next(iter(self._attr_supported_color_modes))
+
     async def async_turn_on(self, **kwargs: Any) -> None:
-        LOGGER.debug("async_turn_on called")
-        LOGGER.debug("kwargs: %s", kwargs)
-        LOGGER.debug(f"Current state: color_mode={self._instance.color_mode}, effect={self._instance.effect}, hs_color={self._instance.hs_color}, brightness={self._instance.brightness}")
+        """Turn the light on."""
+        _LOGGER.debug("turn_on called with kwargs: %s", kwargs)
 
-        if not self.is_on:
-            await self._instance.turn_on()
+        # Ensure light is on
+        if not self._device.is_on:
+            await self._device.turn_on()
 
-        on_brightness = kwargs.get(ATTR_BRIGHTNESS)
-        if on_brightness is None and self._instance.brightness is not None and self._instance.brightness > 0:
-            on_brightness = self._instance.brightness
-        elif on_brightness is None or on_brightness == 0:
-            on_brightness = 255
+        # Determine brightness
+        brightness = kwargs.get(ATTR_BRIGHTNESS)
+        if brightness is None:
+            brightness = self._device.brightness or 255
 
-
-        if ATTR_COLOR_TEMP_KELVIN not in kwargs and ATTR_HS_COLOR not in kwargs and ATTR_EFFECT not in kwargs and ATTR_RGB_COLOR not in kwargs:
-            # i.e. only a brightness change
-            LOGGER.debug(f"Brightness-only change detected. Current color_mode: {self._instance.color_mode}, effect: {self._instance.effect}")
-            if self._instance.effect is not None and self._instance.effect is not EFFECT_OFF:
-                # Before HA 2024.2
-
-                # effect check go first because of the way HA handles brightness changes
-                # if there is no color mode set, the brightness slider in the UI gets disabled
-                # so we bodge it by keep the color mode set while adjusting the effect brightness.
-
-                #HA 2024.2 changes this, setting color mode to "brightness" should allow to change effects brightness as well as introduces the predefined EFFECT_OFF status
-                kwargs[ATTR_EFFECT] = self._instance.effect
-                LOGGER.debug(f"Added effect to kwargs: {self._instance.effect}")
-            elif self._instance.color_mode is ColorMode.COLOR_TEMP:
-                kwargs[ATTR_COLOR_TEMP_KELVIN] = self._instance.color_temp_kelvin
-                LOGGER.debug(f"Added color_temp to kwargs: {self._instance.color_temp_kelvin}")
-            elif self._instance.color_mode is ColorMode.HS:
-                kwargs[ATTR_HS_COLOR] = self._instance.hs_color
-                LOGGER.debug(f"Added hs_color to kwargs: {self._instance.hs_color}")
-            elif self._instance.color_mode is ColorMode.RGB:
-                kwargs[ATTR_RGB_COLOR] = self._instance.rgb_color
-                LOGGER.debug(f"Added rgb_color to kwargs: {self._instance.rgb_color}")
-
-        if ATTR_BRIGHTNESS in kwargs and ATTR_EFFECT == EFFECT_OFF:
-            self._instance._effect = EFFECT_OFF
-        
-        # Set foreground color/effect
-        if ATTR_COLOR_TEMP_KELVIN in kwargs:
-            LOGGER.debug(f"Setting color temp: {kwargs[ATTR_COLOR_TEMP_KELVIN]}, brightness: {on_brightness}")
-            self._instance._color_mode = ColorMode.COLOR_TEMP
-            await self._instance.set_color_temp_kelvin(kwargs[ATTR_COLOR_TEMP_KELVIN], on_brightness)
-        elif ATTR_HS_COLOR in kwargs:
-            # set_hs_color -> set_color includes background color in the packet
-            LOGGER.debug(f"Setting hs_color: {kwargs[ATTR_HS_COLOR]}, brightness: {on_brightness}")
-            await self._instance.set_hs_color(kwargs[ATTR_HS_COLOR], on_brightness)
-        elif ATTR_RGB_COLOR in kwargs:
-            LOGGER.debug(f"Setting rgb_color: {kwargs[ATTR_RGB_COLOR]}, brightness: {on_brightness}")
-            await self._instance.set_rgb_color(kwargs[ATTR_RGB_COLOR], on_brightness)
-        elif ATTR_EFFECT in kwargs and kwargs[ATTR_EFFECT] != EFFECT_OFF:
+        # Handle effect
+        if ATTR_EFFECT in kwargs:
             effect = kwargs[ATTR_EFFECT]
-            # Don't try to set unknown effects
-            if effect.startswith("Unknown Effect"):
-                LOGGER.warning(f"Cannot set unknown effect: {effect}. Ignoring.")
-            else:
-                LOGGER.debug(f"Setting effect: {effect}, brightness: {on_brightness}")
-                await self._instance.set_effect(effect, on_brightness)
-        else:
-            LOGGER.debug(f"No color/effect attribute to set. kwargs keys: {kwargs.keys()}")
-        
-        self.async_write_ha_state()
+            if effect:
+                await self._device.set_effect(effect)
+                return
+
+        # Handle color temperature
+        if ATTR_COLOR_TEMP_KELVIN in kwargs:
+            await self._device.set_color_temp(kwargs[ATTR_COLOR_TEMP_KELVIN], brightness)
+            return
+
+        # Handle RGB color
+        if ATTR_RGB_COLOR in kwargs:
+            await self._device.set_rgb_color(kwargs[ATTR_RGB_COLOR], brightness)
+            return
+
+        # Just brightness change - resend current color/mode
+        # IMPORTANT: Check effect FIRST since it takes priority over stored color values
+        if ATTR_BRIGHTNESS in kwargs:
+            if self._device.effect:
+                # Re-send effect with new brightness
+                await self._device.set_effect(
+                    self._device.effect,
+                    speed=self._device.effect_speed,
+                    brightness=brightness
+                )
+            elif self._device.color_temp_kelvin and self._device.has_color_temp:
+                await self._device.set_color_temp(
+                    self._device.color_temp_kelvin, brightness
+                )
+            elif self._device.rgb_color and self._device.has_rgb:
+                await self._device.set_rgb_color(self._device.rgb_color, brightness)
 
     async def async_turn_off(self, **kwargs: Any) -> None:
-        await self._instance.turn_off()
-        self.async_write_ha_state()
-
-    async def async_update(self) -> None:
-        LOGGER.debug("async update called")
-        await self._instance.update()
-    
-    def light_local_callback(self):
-        LOGGER.debug(f"light_local_callback triggered for {self._instance._mac}")
-        LOGGER.debug(f"Current state: is_on={self._instance.is_on}, brightness={self._instance.brightness}")
-        self.schedule_update_ha_state()
-
-    def update_ha_state(self) -> None:
-        LOGGER.debug("update_ha_state called")
-        if self.hs_color is None and self.color_temp_kelvin is None and self.rgb_color is None:
-            if self._instance._effect is not None and self._instance._effect is not EFFECT_OFF:
-                self._color_mode = ColorMode.BRIGHTNESS 
-                #2024.2 We can use brightness color mode so even when we don't know the state of the light the brightness can be controlled
-                #2025.3 ColorMode.BRIGHTNESS is not considered a valid supported color when on standalone, this gets ignored when
-                #light is rendering an effect 
-            else:
-                self._color_mode = ColorMode.UNKNOWN
-                #2025.3 When not sure of color mode ColorMode.UNKNOWN avoids throwing errors on unsupported combination of color modes
-        elif self.hs_color is not None:
-            self._color_mode = ColorMode.HS
-        # elif self.rgb_color is not None:
-        #     self._color_mode = ColorMode.RGB
-        elif self.color_temp_kelvin is not None:
-            self._color_mode = ColorMode.COLOR_TEMP
-        self.available = self._instance.is_on is not None
-        self.async_write_ha_state()
+        """Turn the light off."""
+        await self._device.turn_off()
 
 
-class LEDNETWFBackgroundLight(LightEntity):
-    """Background light entity for devices that support background color (0x56, 0x80)."""
+class LEDNetWFBackgroundLight(LightEntity):
+    """Background light entity for devices that support background color (0x56, 0x80).
+
+    This entity allows setting the background color for static effects (2-10).
+    The background color is only visible when running a compatible static effect.
+    """
+
     _attr_has_entity_name = True
+    _attr_name = "Background"
+    _attr_icon = "mdi:layers-triple"
 
-    def __init__(
-        self, lednetwfinstance: LEDNETWFInstance, name: str, entry_id: str
-    ) -> None:
-        self._instance = lednetwfinstance
-        self._entry_id = entry_id
-        self._attr_supported_color_modes = {ColorMode.HS}
-        self._attr_name = "Background"  # This will be the entity name suffix
-        self._attr_unique_id = f"{self._instance.mac}_background"
-        self._device_name = name  # Store the device name for reference
-        self._is_on = True  # Track on/off state independently from brightness
-        # Register this entity's callback
-        self._instance.register_callback(self.bg_light_local_callback)
-        
-    @property
-    def available(self):
-        """Background light is only available for static effects (excluding Solid Color)."""
-        if self._instance.is_on is None:
-            return False
-        
-        current_effect = self._instance.effect
-        
-        # Only available for static effects (Static Effect 2-10)
-        # NOT available for: color mode, Solid Color, dynamic effects, or sound reactive
-        if current_effect is None or current_effect == EFFECT_OFF:
-            return False
-            
-        # Static effects (but not "Solid Color" which is now EFFECT_OFF)
-        return "Static Effect" in current_effect
+    def __init__(self, device: LEDNetWFDevice, entry: ConfigEntry) -> None:
+        """Initialize the background light entity."""
+        self._device = device
+        self._entry = entry
 
-    @property
-    def brightness(self):
-        return self._instance.bg_brightness
-    
-    @property
-    def brightness_step_pct(self):
-        return 10
-    
-    @property
-    def is_on(self) -> Optional[bool]:
-        # Background light has its own on/off state, independent of brightness value
-        return self._instance.is_on and self._is_on
+        # Set up unique ID
+        self._attr_unique_id = f"{device.address}_background"
+
+        # Background light only supports RGB
+        self._attr_supported_color_modes = {ColorMode.RGB}
+        self._attr_color_mode = ColorMode.RGB
+
+        # No effects for background light
+        self._attr_supported_features = LightEntityFeature(0)
+
+        # Track independent on/off state (brightness 0 = off)
+        self._is_on = True
+
+        # Register callback for state updates
+        device.register_callback(self._handle_state_update)
+
+    async def async_will_remove_from_hass(self) -> None:
+        """Handle entity removal."""
+        self._device.unregister_callback(self._handle_state_update)
+
+    @callback
+    def _handle_state_update(self) -> None:
+        """Handle state updates from the device."""
+        self.async_write_ha_state()
 
     @property
-    def supported_color_modes(self) -> set[ColorMode]:
-        """Flag supported color modes."""
-        return self._attr_supported_color_modes
-
-    @property
-    def hs_color(self):
-        """Return the hs color value."""
-        return self._instance.bg_hs_color
-  
-    @property
-    def rgb_color(self):
-        """Return the rgb color value."""
-        return self._instance.bg_rgb_color
-
-    @property
-    def color_mode(self):
-        """Return the color mode of the light."""
-        return ColorMode.HS
-    
-    @property
-    def firmware_version(self):
-        return self._instance.firmware_version
-    
-    @property
-    def device_info(self):
+    def device_info(self) -> DeviceInfo:
         """Return device info - links to same device as main light."""
         return DeviceInfo(
-            identifiers={
-                (DOMAIN, self._instance.mac)
-            },
+            identifiers={(DOMAIN, self._device.address)},
             # Don't set name here - it will link to existing device with same identifiers
-            connections={(device_registry.CONNECTION_NETWORK_MAC, self._instance.mac)},
-            model=f"0x{self._instance.model_number:02x}",
-            sw_version=self.firmware_version
         )
 
     @property
-    def should_poll(self):
-        return False
+    def available(self) -> bool:
+        """Return True if entity is available.
+
+        Background light is only available when running a static effect (2-10).
+        Not available for: color mode, solid color, dynamic effects, or sound reactive.
+        """
+        if self._device.is_on is None:
+            return False
+
+        # Must be running a static effect
+        return self._device.is_bg_color_available()
 
     @property
-    def name(self) -> str:
-        return self._attr_name
-    
+    def is_on(self) -> bool | None:
+        """Return True if background light is on.
+
+        Background light has its own on/off state, independent of brightness value.
+        """
+        return self._device.is_on and self._is_on
+
     @property
-    def icon(self):
-        return "mdi:layers-triple"
-    
+    def brightness(self) -> int | None:
+        """Return the background brightness."""
+        return self._device.bg_brightness
+
+    @property
+    def rgb_color(self) -> tuple[int, int, int] | None:
+        """Return background RGB color."""
+        return self._device.bg_rgb_color
+
     async def async_turn_on(self, **kwargs: Any) -> None:
-        LOGGER.debug("Background light async_turn_on called")
-        LOGGER.debug("kwargs: %s", kwargs)
+        """Turn the background light on."""
+        _LOGGER.debug("Background light turn_on called with kwargs: %s", kwargs)
 
         # Ensure main light is on
-        if not self._instance.is_on:
-            await self._instance.turn_on()
+        if not self._device.is_on:
+            await self._device.turn_on()
 
         # Mark as on
         self._is_on = True
 
-        # Determine brightness - preserve current brightness if not explicitly set
-        on_brightness = kwargs.get(ATTR_BRIGHTNESS)
-        if on_brightness is None:
-            # If no brightness is specified, use current brightness if reasonable
-            # Threshold of 10 to avoid very dim colors when turning on
-            if self._instance.bg_brightness is not None and self._instance.bg_brightness >= 10:
-                on_brightness = self._instance.bg_brightness
+        # Determine brightness
+        brightness = kwargs.get(ATTR_BRIGHTNESS)
+        if brightness is None:
+            # Preserve current brightness if reasonable
+            if self._device.bg_brightness >= 10:
+                brightness = self._device.bg_brightness
             else:
-                # Check if we have a saved brightness from before turning off
-                if hasattr(self, '_last_brightness') and self._last_brightness >= 10:
-                    on_brightness = self._last_brightness
-                    LOGGER.debug(f"Restoring last brightness: {on_brightness}")
-                else:
-                    # Default to full brightness when turning on from off/very dim state
-                    on_brightness = 255
-                    LOGGER.debug(f"Background brightness was {self._instance.bg_brightness}, defaulting to 255")
+                # Default to foreground brightness (user expectation: BG matches FG)
+                brightness = self._device.brightness or 255
 
-        if ATTR_HS_COLOR in kwargs:
-            await self._instance.set_bg_hs_color(kwargs[ATTR_HS_COLOR], on_brightness)
-        elif ATTR_RGB_COLOR in kwargs:
-            # Convert RGB to HS
-            from homeassistant.util.color import color_RGB_to_hs
-            hs = color_RGB_to_hs(*kwargs[ATTR_RGB_COLOR])
-            await self._instance.set_bg_hs_color(hs, on_brightness)
+        # Handle RGB color
+        if ATTR_RGB_COLOR in kwargs:
+            rgb = kwargs[ATTR_RGB_COLOR]
         else:
-            # Just turning on or brightness change - preserve color
-            await self._instance.set_bg_hs_color(self._instance.bg_hs_color, on_brightness)
-        
-        self.async_write_ha_state()
+            # Use current background color or default to white
+            rgb = self._device.bg_rgb_color or (255, 255, 255)
+
+        await self._device.set_bg_color(rgb, brightness)
 
     async def async_turn_off(self, **kwargs: Any) -> None:
-        # Save current brightness before turning off (if it's reasonable)
-        if self._instance.bg_brightness is not None and self._instance.bg_brightness >= 10:
-            self._last_brightness = self._instance.bg_brightness
-            LOGGER.debug(f"Saving brightness {self._last_brightness} before turn off")
+        """Turn the background light off.
 
-        # Mark as off and set background to black (brightness 0)
+        Sets brightness to 0 to hide the background color.
+        """
+        _LOGGER.debug("Background light turn_off called")
+
+        # Mark as off
         self._is_on = False
-        await self._instance.set_bg_hs_color(self._instance.bg_hs_color, 0)
-        self.async_write_ha_state()
 
-    async def async_update(self) -> None:
-        LOGGER.debug("Background light async_update called")
-        await self._instance.update()
-        self.async_write_ha_state()
-    
-    def bg_light_local_callback(self):
-        """Called when device state changes via notification."""
-        self.async_write_ha_state()
+        # Set background to black (brightness 0)
+        rgb = self._device.bg_rgb_color or (255, 255, 255)
+        await self._device.set_bg_color(rgb, 0)
