+++ custom_components/lednetwf_ble/number.py
@@ -1,75 +1,86 @@
+"""Number platform for LEDnetWF BLE v2 integration."""
 from __future__ import annotations
-from homeassistant.components.number import (
-    NumberEntity,
-    NumberEntityDescription,
-)
-from .lednetwf import LEDNETWFInstance
-from .const import DOMAIN
-from homeassistant.helpers.entity import DeviceInfo
-from homeassistant.helpers.entity_platform import AddEntitiesCallback
-from homeassistant.helpers import device_registry
-from homeassistant.core import HomeAssistant, callback
-from homeassistant.config_entries import ConfigEntry
+
 import logging
+
+from homeassistant.components.number import NumberEntity, NumberMode
+from homeassistant.config_entries import ConfigEntry
+from homeassistant.core import HomeAssistant, callback
+from homeassistant.helpers.device_registry import DeviceInfo
+from homeassistant.helpers.entity_platform import AddEntitiesCallback
+
+from .const import DOMAIN, DEFAULT_EFFECT_SPEED
+from .device import LEDNetWFDevice
 
 _LOGGER = logging.getLogger(__name__)
 
+
 async def async_setup_entry(
     hass: HomeAssistant,
-    config_entry: ConfigEntry,
+    entry: ConfigEntry,
     async_add_entities: AddEntitiesCallback,
 ) -> None:
-    instance = hass.data[DOMAIN][config_entry.entry_id]
-    await instance.update()
-    async_add_entities([LEDNETWFSpeedSlider(instance, "Effect speed", config_entry.entry_id)])
+    """Set up the number platform."""
+    device: LEDNetWFDevice = hass.data[DOMAIN][entry.entry_id]
 
-class LEDNETWFSpeedSlider(NumberEntity):
-    """LEDNETWF Slider for effect speed."""
+    entities: list[NumberEntity] = []
 
-    def __init__(self, lednetfInstance: LEDNETWFInstance, attr_name: str, entry_id: str) -> None:
-        self._instance             = lednetfInstance
-        self._attr_has_entity_name = True
-        #self._attr_translation_key = attr_name # Can't get this to work
-        self._attr_name            = attr_name
-        self._attr_unique_id       = self._instance.mac
-        
-        # Register for state updates
-        self._instance.register_callback(self.speed_local_callback)
+    # Only add effect speed if device supports effects
+    if device.has_effects:
+        entities.append(LEDNetWFEffectSpeed(device))
 
-    @property
-    def available(self):
-        return self._instance.is_on is not None
+    if entities:
+        async_add_entities(entities)
 
-    @property
-    def name(self) -> str:
-        return self._attr_name
 
-    @property
-    def unique_id(self) -> str:
-        """Return the unique id."""
-        return self._attr_unique_id
+class LEDNetWFEffectSpeed(NumberEntity):
+    """Effect speed control for LEDnetWF devices.
 
-    @property
-    def native_value(self) -> int | None:
-        # Always read from model to get latest state
-        return self._instance._model_interface.effect_speed
-    
+    This can serve as "speed" for effects or "sensitivity" for sound-reactive modes.
+    """
+
+    _attr_has_entity_name = True
+    _attr_name = "Effect Speed"
+    _attr_native_min_value = 0
+    _attr_native_max_value = 100
+    _attr_native_step = 1
+    _attr_mode = NumberMode.SLIDER
+    _attr_icon = "mdi:speedometer"
+
+    def __init__(self, device: LEDNetWFDevice) -> None:
+        """Initialize the number entity."""
+        self._device = device
+        self._attr_unique_id = f"{device.address}_effect_speed"
+
+        # Register callback for state updates
+        device.register_callback(self._handle_state_update)
+
+    async def async_will_remove_from_hass(self) -> None:
+        """Handle entity removal."""
+        self._device.unregister_callback(self._handle_state_update)
+
     @callback
-    def speed_local_callback(self) -> None:
-        """Handle state updates from device."""
-        _LOGGER.debug("Speed callback triggered, effect_speed=%s", 
-                     self._instance._model_interface.effect_speed)
-        self.schedule_update_ha_state()
+    def _handle_state_update(self) -> None:
+        """Handle state updates from the device."""
+        self.async_write_ha_state()
 
     @property
-    def device_info(self):
+    def device_info(self) -> DeviceInfo:
         """Return device info."""
         return DeviceInfo(
-            identifiers={(DOMAIN, self._instance.mac)},
-            connections={(device_registry.CONNECTION_NETWORK_MAC,
-                          self._instance.mac)},
+            identifiers={(DOMAIN, self._device.address)},
         )
 
+    @property
+    def available(self) -> bool:
+        """Return True if entity is available."""
+        return self._device.is_on is not None
+
+    @property
+    def native_value(self) -> float:
+        """Return the current effect speed."""
+        return self._device.effect_speed
+
     async def async_set_native_value(self, value: float) -> None:
-        """Update the current value."""
-        await self._instance.set_effect_speed(int(value))
+        """Set the effect speed."""
+        await self._device.set_effect_speed(int(value))
