+++ custom_components/lednetwf_ble/strings.json
