--- custom_components/lednetwf_ble_3/manifest.json
@@ -1,31 +1,0 @@
-{
-    "domain": "lednetwf_ble",
-    "name": "LEDnetWF BLE v2",
-    "bluetooth": [
-        {
-            "local_name": "LEDnetWF*"
-        },
-        {
-            "local_name": "IOTWF*"
-        },
-        {
-            "local_name": "IOTB*"
-        }
-    ],
-    "codeowners": [
-        "@8none1"
-    ],
-    "config_flow": true,
-    "dependencies": [
-        "bluetooth"
-    ],
-    "documentation": "https://github.com/8none1/lednetwf_ble",
-    "integration_type": "device",
-    "iot_class": "local_push",
-    "issue_tracker": "https://github.com/8none1/lednetwf_ble/issues",
-    "requirements": [
-        "bleak-retry-connector>=1.17.1",
-        "bleak>=0.17.0"
-    ],
-    "version": "2.0.0"
-}
